Remove folder and (most) OS icons, clean up colors
Now, the OS icon is only rendered when in a WSL environment as it isn't
that useful or easy to read the rest of the time.

The folder icon was removed from the path part of the prompt because it,
too, was hard to read and served no helpful purpose.

Colors were cleaned up so that names are used everywhere and unused bits
were removed.